Moved IServiceProxy and IHostServiceProxy to Coral.PluginBase. Removed Coral.PluginHost dependency from Coral.Plugin.LastFM
