+++ src/Coral.Plugin.LastFM/LastFmController.cs
@@ -1,11 +1,5 @@
 ﻿using Microsoft.AspNetCore.Mvc;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using Coral.PluginBase;
-using Coral.PluginHost;
 
 namespace Coral.Plugin.LastFM
 {
