+++ src/Coral.Plugin.LastFM/LastFmService.cs
@@ -4,19 +4,13 @@ using Coral.Dto.Models;
 using Coral.Events;
 using Coral.Plugin.LastFM.ResponseTypes;
 using Coral.PluginBase;
-using Coral.PluginHost;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
 using RestSharp;
 using RestSharp.Serializers.Json;
-using System;
-using System.Collections.Generic;
-using System.Linq;
 using System.Security.Cryptography;
 using System.Text;
 using System.Text.Json;
-using System.Text.Json.Serialization;
-using System.Threading.Tasks;
 
 namespace Coral.Plugin.LastFM
 {
