+++ src/Coral.PluginBase/IHostServiceProxy.cs
@@ -1,0 +1,7 @@
+﻿namespace Coral.PluginBase;
+
+public interface IHostServiceProxy
+{
+    public TType GetHostService<TType>()
+        where TType : class;
+}
