+++ src/Coral.PluginBase/IServiceProxy.cs
@@ -1,0 +1,7 @@
+﻿namespace Coral.PluginBase;
+
+public interface IServiceProxy
+{
+    public TType GetService<TType>()
+        where TType : class;
+}
